Removed unnecessary casts from macros.

+++ include/common.h
@@ -26,7 +26,7 @@ extern "C" {
  * \pre -70776 <= n && n < 70777.
  */
 #define AMARU_LOG10_POW2(n) \
-  ((int32_t) ((((uint64_t) 1292913986) * ((int32_t) n)) >> 32))
+  ((int32_t) ((((uint64_t) 1292913986) * (n)) >> 32))
 
 /**
  * \brief Returns the integer part of log_2(10^n).
@@ -34,7 +34,7 @@ extern "C" {
  * \pre -55266 <= n && n < 55267.
  */
 #define AMARU_LOG2_POW10(n) \
-  ((int32_t) ((((uint64_t) 3566893131) * ((int32_t) n)) >> 30))
+  ((int32_t) ((((uint64_t) 3566893131) * (n)) >> 30))
 
 /**
  * \brief Returns the integer part of log_5(2^n).
@@ -42,7 +42,7 @@ extern "C" {
  * \pre -227267 <= n && n < 227268.
  */
 #define AMARU_LOG5_POW2(n) \
-  ((int32_t) ((((uint64_t) 1849741732) * ((int32_t) n)) >> 32))
+  ((int32_t) ((((uint64_t) 1849741732) * (n)) >> 32))
 
 #ifdef __cplusplus
 }
